Refactor also chat overview components
